+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
-# Bug fixes
+# General Info
 
-- pahe: Fix issue with alleged episode count exceeding actual number of episodes
+- Add termux port, [senpcli](https://github.com/SenZmaKi/Senpwai/blob/master/docs/senpcli-guide.md) in now installable on android using [termux](https://github.com/termux/termux-app)
+- Fix Gogoanime by updating domain name to anitaku.so